- Get rid of warnings in 64-bit compilers on ERR_POST_X and LOG_POST_X macros. - Revert change of NCBI_NAME2 macro: some code rely on fact that macros doesn't expand here. JIRA: CXX-3
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@35730 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/corelib/ncbidiag.hpp
@@ -337,12 +337,12 @@ inline void CheckErrSubcodeX(int)
 /// @sa LOG_POST_X, ERR_POST_X
 #define NCBI_CHECK_ERR_SUBCODE_X_NAME(name, subcode)                  \
     NCBI_NS_NCBI::CheckErrSubcodeX(                                   \
-        sizeof(NCBI_NS_NCBI::WRONG_ERROR_SUBCODE_IN_POST_MACRO<       \
+        (int)sizeof(NCBI_NS_NCBI::WRONG_ERROR_SUBCODE_IN_POST_MACRO<  \
               NCBI_ERRCODE_X_NAME(name), subcode,                     \
               NCBI_MAX_ERR_SUBCODE_X_NAME(name),                      \
               ((unsigned int)subcode >                                \
                     (unsigned int)NCBI_MAX_ERR_SUBCODE_X_NAME(name))  \
-                                                              >)      \
+                                                                   >) \
                                   )
 
 #endif  // if defined(NCBI_COMPILER_GCC) && NCBI_COMPILER_VERSION < 350 else
